moves list/items to a list page, simplifies index page, configures webpack to bundle js per page
