+++ index.html
@@ -24,22 +24,11 @@
     <nav>
       <div class="navContents">
         <p>Pantry Pal</p>
-        <button id="openOptionsModalButton">Settings</button>
       </div>
     </nav>
     <div class="container">
-      <!-- ADD LIST/ITEM MODAL TRIGGERS -->
+      <!-- ADD LIST MODAL TRIGGER -->
       <div class="openModalButtonContainer">
-        <button
-          class="openModalButton"
-          id="openModalButton"
-          aria-haspopup="dialog"
-          aria-controls="modal"
-        >
-          <div class="plus"></div>
-          <p>Add Item</p>
-        </button>
-
         <button
           class="openModalButton"
           id="openAddListModalButton"
@@ -51,79 +40,6 @@
         </button>
       </div>
 
-      <!-- MODAL STRUCTURE -->
-      <!-- ADD ITEM MODAL -->
-      <div
-        id="modal"
-        class="modal"
-        role="dialog"
-        aria-labelledby="modalTitle"
-        aria-modal="true"
-        aria-hidden="true"
-      >
-        <div class="modal-content">
-          <button id="closeModal" class="close" aria-label="Close">
-            &times;
-          </button>
-          <h2 id="modalTitle">Add New Item</h2>
-          <form id="addItemForm" novalidate>
-            <label for="nameInput">
-              Item
-              <input
-                type="text"
-                id="nameInput"
-                placeholder="Grocery Item"
-                aria-describedby="nameError"
-                required
-              />
-              <span class="error" id="nameError"></span>
-            </label>
-
-            <div class="formInputGroup">
-              <label for="quantityInput">
-                Quantity
-                <input type="number" id="quantityInput" value="1" />
-              </label>
-
-              <label for="quantityUnitInput">
-                Unit
-                <select name="quantityUnitInput" id="quantityUnitInput">
-                  <!-- GENERATED IN FORM.JS AND OPTIONSDATA.JS -->
-                </select>
-              </label>
-            </div>
-
-            <!-- <label for="priceInput">
-              Price
-              <input
-                id="priceInput"
-                type="number"
-                placeholder="1.99"
-                step="0.01"
-                min="0"
-                aria-describedby="priceError"
-              />
-              <span class="error" id="priceError"></span>
-            </label> -->
-
-            <label for="sectionInput">
-              Store Section
-              <select name="sectionInput" id="sectionInput">
-                <!-- GENERATED IN FORM.TS AND OPTIONSDATA.JS -->
-              </select>
-            </label>
-
-            <div class="formActionsDiv">
-              <button type="submit" class="addButton" id="addItemButton">
-                Add Item
-              </button>
-              <button id="cancelButton" class="cancelButton" type="button">
-                Cancel
-              </button>
-            </div>
-          </form>
-        </div>
-      </div>
       <!-- ADD LIST MODAL -->
       <div
         id="addListModal"
@@ -169,102 +85,13 @@
     </div>
 
     <!-- FILTERS AND OPTIONS SIDEBAR -->
-    <div
-      id="optionsModal"
-      class="optionsModal"
-      role="dialog"
-      aria-labelledby="optionsModalTitle"
-      aria-modal="true"
-      aria-hidden="true"
-    >
-      <div class="optionsSettingsContainer">
-        <div class="optionsModalTitleContainer">
-          <h2 class="optionsModalTitle">Settings</h2>
-          <button
-            id="closeOptionsModal"
-            class="close"
-            aria-label="Close Settings Panel"
-          >
-            &times;
-          </button>
-        </div>
+    <!-- TODO: We'll want something like the items filters for lists -->
 
-        <div class="filterInputs">
-          <div class="filterControl">
-            <p id="autoSortLabel" class="controlText">Auto Sort</p>
-            <label class="switch">
-              <input
-                name="autoSort"
-                aria-labelledby="autoSortLabel"
-                id="autoSort"
-                type="checkbox"
-                class="switch-input"
-              />
-              <span class="switch-label" data-on="On" data-off="Off"></span>
-              <span class="switch-handle"></span>
-            </label>
-          </div>
+    <!-- LISTS -->
+    <ul id="listsDiv">
+      <!-- RENDERED IN DOMUTILS  -->
+    </ul>
 
-          <div class="filterControl">
-            <p id="hideCheckedLabel" class="controlText">Hide Checked</p>
-            <label class="switch">
-              <input
-                id="hideChecked"
-                aria-labelledby="hideCheckedLabel"
-                type="checkbox"
-                class="switch-input"
-              />
-              <span class="switch-label" data-on="On" data-off="Off"></span>
-              <span class="switch-handle"></span>
-            </label>
-          </div>
-
-          <div class="filterControl">
-            <p id="sectionSortLabel" class="controlText">Sort by Section</p>
-            <label class="switch">
-              <input
-                id="sectionSort"
-                aria-labelledby="sectionSortLabel"
-                type="checkbox"
-                class="switch-input"
-              />
-              <span class="switch-label" data-on="On" data-off="Off"></span>
-              <span class="switch-handle"></span>
-            </label>
-          </div>
-        </div>
-
-        <div class="optionsModalFooter">
-          <button id="backupData" class="downloadDataButton">Download</button>
-
-          <div class="uploadBackupArea">
-            <form id="uploadBackupForm">
-              <label class="fileInputLabel" for="backupFileInput">
-                <input
-                  type="file"
-                  id="backupFileInput"
-                  accept="application/json"
-                />
-              </label>
-              <button class="uploadBackupButton" type="submit">Upload</button>
-              <!-- <div id="spinner">Loading...</div> -->
-            </form>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <!-- SHOPPING ITEMS LIST -->
-    <ul id="itemsDiv"></ul>
-
-    <!-- SECTION SORTING -->
-    <div class="stickyQuickSortFooter" id="stickyQuickSortFooter">
-      <div id="quickSortWrapper">
-        <div id="quickSortDiv">
-          <!-- CONTENT RENDERED IN DOMUTILS.TS -->
-        </div>
-      </div>
-    </div>
     <script src="./index.js"></script>
   </body>
 </html>
