+++ list.html
@@ -39,16 +39,6 @@
           <div class="plus"></div>
           <p>Add Item</p>
         </button>
-
-        <button
-          class="openModalButton"
-          id="openAddListModalButton"
-          aria-haspopup="dialog"
-          aria-controls="addListModal"
-        >
-          <div class="plus"></div>
-          <p>Add List</p>
-        </button>
       </div>
 
       <!-- MODAL STRUCTURE -->
@@ -118,48 +108,6 @@
                 Add Item
               </button>
               <button id="cancelButton" class="cancelButton" type="button">
-                Cancel
-              </button>
-            </div>
-          </form>
-        </div>
-      </div>
-      <!-- ADD LIST MODAL -->
-      <div
-        id="addListModal"
-        class="listModal"
-        role="dialog"
-        aria-labelledby="addListModalTitle"
-        aria-modal="true"
-        aria-hidden="true"
-      >
-        <div class="addListModalContent">
-          <button id="closeListModal" class="close" aria-label="Close">
-            &times;
-          </button>
-          <h2 class="addListModalTitle">Add a List</h2>
-          <form id="addListForm" novalidate>
-            <label for="listNameInput">
-              Item
-              <input
-                type="text"
-                id="listNameInput"
-                placeholder="Taco Tuesday"
-                aria-describedby="listNameError"
-                required
-              />
-              <span class="error" id="listNameError"></span>
-            </label>
-
-            <div class="formActionsDiv">
-              <button type="submit" class="addButton" id="addListItemButton">
-                Add List
-              </button>
-              <button
-                id="cancelAddListButton"
-                class="cancelButton"
-                type="button"
-              >
                 Cancel
               </button>
             </div>
@@ -265,6 +213,6 @@
         </div>
       </div>
     </div>
-    <script src="./index.js"></script>
+    <script src="./list.js"></script>
   </body>
 </html>
