+++ styles/buttons.css
@@ -19,6 +19,7 @@
 }
 
 button {
+  cursor: pointer;
   border: none;
   border-radius: 8px;
   color: var(--button-text-color);
